新增管理页面的接口

+++ src/main/java/com/example/demo/domain/po/UserPO.java
@@ -11,5 +11,5 @@ public class UserPO {
     String username;
     String password;
     String type;
-    Integer companyId;
+    Long companyId;
 }
