+++ src/main/java/com/example/demo/repository/impl/CompanyRepositoryImpl.java
@@ -13,7 +13,7 @@ public class CompanyRepositoryImpl {
         return companyRepository.save(companyPO);
     }
 
-    CompanyPO find(Integer companyId) {
+    CompanyPO find(Long companyId) {
         return companyRepository.find(companyId);
     }
